backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pandas==2.1.3
lxml==4.9.3
scikit-learn==1.3.2
numpy==1.24.3
openai>=1.54.0
python-dotenv==1.0.0

